app/backend/requirements.txt
# Runtime
Django>=4.2,<5
djangorestframework>=3.14,<4
gunicorn>=20.1,<21
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
dj-database-url>=1.0,<2
django-cors-headers>=4.0,<5
whitenoise>=6.5,<7
Pillow>=9.0,<10

# Optional (task queue / broker)
celery>=5.3,<6
redis>=4.5,<5

# Optional (migrations / SQL toolkit)
alembic>=1.11,<2
SQLAlchemy>=1.4,<2

# Dev / Testing
fastapi>=0.100,<1
uvicorn>=0.22,<1
pydantic>=2.9,<3
pytest>=7.0,<8
pytest-django>=4.5,<5
black>=23.0,<24
isort>=5.10,<6
flake8>=5.0,<6